[UPDATE] Image weather to animation [ADD] Forecast

+++ app/src/main/java/com/lms/weatherapp/model/common/Data.kt
@@ -6,14 +6,14 @@ import com.squareup.moshi.JsonClass
 
 @JsonClass(generateAdapter = true)
 data class Metric(
-    @SerializedName("Value")val value : Int,
+    @SerializedName("Value")val value : Double,
     @SerializedName("Unit") val unit: String,
     @SerializedName("UnitType") val unitType: Int,
 )
 
 @JsonClass(generateAdapter = true)
 data class Imperial(
-    @SerializedName("Value")val value : Int,
+    @SerializedName("Value")val value : Double,
     @SerializedName("Unit") val unit: String,
     @SerializedName("UnitType") val unitType: Int,
 )
